chore(llm): rename Analytics/constants to Analytics/enums
