+++ apps/ledger-live-mobile/src/newArch/features/Analytics/enums.ts
