+++ apps/ledger-live-mobile/src/newArch/features/WalletSync/hooks/useAddMember.ts
@@ -15,7 +15,7 @@ import { TrustchainResult, TrustchainResultType } from "@ledgerhq/ledger-key-rin
 import { useCallback, useRef } from "react";
 import { Device } from "@ledgerhq/live-common/hw/actions/types";
 import { useNavigation } from "@react-navigation/native";
-import { AnalyticsEvents } from "LLM/features/Analytics/constants";
+import { AnalyticsEvents } from "~/newArch/features/Analytics/enums";
 import { track } from "~/analytics";
 import { WalletSyncNavigatorStackParamList } from "~/components/RootNavigator/types/WalletSyncNavigator";
 import { StackNavigatorNavigation } from "~/components/RootNavigator/types/helpers";
