+++ apps/ledger-live-mobile/src/newArch/features/WalletSync/hooks/useDestroyTrustchain.ts
@@ -6,7 +6,7 @@ import {
   memberCredentialsSelector,
 } from "@ledgerhq/ledger-key-ring-protocol/store";
 import { useMutation } from "@tanstack/react-query";
-import { AnalyticsEvents } from "LLM/features/Analytics/constants";
+import { AnalyticsEvents } from "~/newArch/features/Analytics/enums";
 import { track } from "~/analytics";
 import { QueryKey } from "./type.hooks";
 import { useCloudSyncSDK } from "./useWatchWalletSync";
