+++ apps/ledger-live-mobile/src/newArch/features/WalletSync/hooks/useQRCodeHost.ts
@@ -12,7 +12,7 @@ import {
   memberCredentialsSelector,
   setTrustchain,
 } from "@ledgerhq/ledger-key-ring-protocol/store";
-import { AnalyticsEvents } from "LLM/features/Analytics/constants";
+import { AnalyticsEvents } from "~/newArch/features/Analytics/enums";
 import { track } from "~/analytics";
 import { useTrustchainSdk } from "./useTrustchainSdk";
 import { Options, Steps } from "../types/Activation";
